Set store_data to save introspection data
Sets the store_data to database to allow introspection data to be
saved and retrieved.  This uses sqlalchamy as the default store.

Note that the database support required an ironic-inspector fix -
https://review.openstack.org/#/c/636692/, which is currently not
in current-tripleo.  This temporarily changes the repo to use
current instead of current-tripleo.  This change will be reverted
once the fix is available.