Wait for PROVISIONING_INTERFACE to be up
We've seen some issues where the static-ip-refresh doesn't happen quickly
enough, and/or static-ip-set from initContainers expires before we start
the ironic containers.

To ensure we don't fail in that case, wait for the provisioning IP to be
assigned, then the exact order of container start is not important, and
we'll wait until static-ip-refresh has configured the nic.