Make searching for Provisioning IP robust
Using grep with a user-supplied query that is likely to contain .
characters is not the most robust method of searching for a string in
some output. For example, passing PROVISIONING_IP=192.168.1.2 would also
match e.g. 192.168.122.1. Avoid this by searching for a fixed string,
including the characters before and after the IP address.

The best way to do this would be using the json output from the ip
command:

    [ -n $(ip -j addr show | jq -r ".[] | .addr_info[] | .local | select(. == ${IRONIC_IP})") ]

and something similar to get the IP address when only the provisioning
interface is specfied:

    ip -j add show scope global up dev ${PROVISIONING_INTERFACE} | jq -r .[0].addr_info[0].local

but continuing to do it with grep avoids having to install the jq tool.